Link bus routes that loop back at the end of the run
Some bus trips are represented in GTFS as ending, and then a new trip
of the same route continues from the same location. This is common for
bi-directional routes or loops. We can use the block ID for these trips
to make sure gaps aren’t shown in our maps.